Fix invalid memory access in runtime test; Fix device tests not running
