+++ wrappers/golang/runtime/tests/device_test_darwin.go
@@ -23,16 +23,6 @@ import (
 	"github.com/stretchr/testify/suite"
 )
 
-func testGetDeviceType(suite *suite.Suite) {
-	expectedDeviceName := "test"
-	config := icicle_runtime.CreateDevice(expectedDeviceName, 0)
-	suite.Equal(expectedDeviceName, config.GetDeviceType())
-
-	expectedDeviceNameLong := "testtesttesttesttesttesttesttesttesttesttesttesttesttesttesttest"
-	configLargeName := icicle_runtime.CreateDevice(expectedDeviceNameLong, 1)
-	suite.NotEqual(expectedDeviceNameLong, configLargeName.GetDeviceType())
-}
-
 func testIsDeviceAvailable(suite *suite.Suite) {
 	test_helpers.ActivateMainDevice()
 	res, err := icicle_runtime.GetDeviceCount()
@@ -47,11 +37,11 @@ func testIsDeviceAvailable(suite *suite.Suite) {
 }
 
 func testSetDefaultDevice(suite *suite.Suite) {
-	test_helpers.ActivateMainDevice()
-
-	activeDevice, err := icicle_runtime.GetActiveDevice()
-	suite.Equal(icicle_runtime.Success, err)
-	suite.Equal(test_helpers.MAIN_DEVICE, *activeDevice)
+	runtime.LockOSThread()
+	defer runtime.UnlockOSThread()
+	
+	defaultDevice := &test_helpers.MAIN_DEVICE
+	icicle_runtime.SetDefaultDevice(defaultDevice)
 
 	outerThreadID := C.get_pthread_id()
 	done := make(chan struct{}, 1)
@@ -70,7 +60,7 @@ func testSetDefaultDevice(suite *suite.Suite) {
 
 		activeDevice, err := icicle_runtime.GetActiveDevice()
 		suite.Equal(icicle_runtime.Success, err)
-		suite.Equal(test_helpers.MAIN_DEVICE, *activeDevice)
+		suite.Equal(defaultDevice, *activeDevice)
 
 		close(done)
 	}()
@@ -78,45 +68,4 @@ func testSetDefaultDevice(suite *suite.Suite) {
 	<-done
 
 	icicle_runtime.SetDefaultDevice(&test_helpers.REFERENCE_DEVICE)
-}
-
-func testRegisteredDevices(suite *suite.Suite) {
-	devices, _ := icicle_runtime.GetRegisteredDevices()
-	suite.Equal([]string{test_helpers.MAIN_DEVICE.GetDeviceType(), "CPU"}, devices)
-}
-
-func testDeviceProperties(suite *suite.Suite) {
-	_, err := icicle_runtime.GetDeviceProperties()
-	suite.Equal(icicle_runtime.Success, err)
-}
-
-func testActiveDevice(suite *suite.Suite) {
-	test_helpers.ActivateMainDevice()
-	activeDevice, err := icicle_runtime.GetActiveDevice()
-	suite.Equal(icicle_runtime.Success, err)
-	suite.Equal(test_helpers.MAIN_DEVICE, *activeDevice)
-	memory1, err := icicle_runtime.GetAvailableMemory()
-	if err == icicle_runtime.ApiNotImplemented {
-		suite.T().Skipf("GetAvailableMemory() function is not implemented on %s device", test_helpers.MAIN_DEVICE.GetDeviceType())
-	}
-	suite.Equal(icicle_runtime.Success, err)
-	suite.Greater(memory1.Total, uint(0))
-	suite.Greater(memory1.Free, uint(0))
-}
-
-type DeviceTestSuite struct {
-	suite.Suite
-}
-
-func (s *DeviceTestSuite) TestGetDeviceType() {
-	s.Run("TestGetDeviceType", test_helpers.TestWrapper(&s.Suite, testGetDeviceType))
-	s.Run("TestIsDeviceAvailable", test_helpers.TestWrapper(&s.Suite, testIsDeviceAvailable))
-	s.Run("TestSetDefaultDevice", test_helpers.TestWrapper(&s.Suite, testSetDefaultDevice))
-	s.Run("TestRegisteredDevices", test_helpers.TestWrapper(&s.Suite, testRegisteredDevices))
-	s.Run("TestDeviceProperties", test_helpers.TestWrapper(&s.Suite, testDeviceProperties))
-	s.Run("TestActiveDevice", test_helpers.TestWrapper(&s.Suite, testActiveDevice))
-}
-
-func TestSuiteDevice(t *testing.T) {
-	suite.Run(t, new(DeviceTestSuite))
 }
