+++ wrappers/golang/runtime/tests/memory_test.go
@@ -11,15 +11,20 @@ import (
 
 func testMalloc(suite *suite.Suite) {
 	test_helpers.ActivateMainDevice()
-	_, err := runtime.Malloc(20)
+	mem, err := runtime.Malloc(20)
 	suite.Equal(runtime.Success, err, "Unable to allocate device memory due to %d", err)
+	runtime.Free(mem)
 }
 
 func testMallocAsync(suite *suite.Suite) {
 	test_helpers.ActivateMainDevice()
 	stream, _ := runtime.CreateStream()
-	_, err := runtime.MallocAsync(20, stream)
+	mem, err := runtime.MallocAsync(20, stream)
 	suite.Equal(runtime.Success, err, "Unable to allocate device memory due to %d", err)
+
+	runtime.FreeAsync(mem, stream)
+	runtime.SynchronizeStream(stream)
+	runtime.DestroyStream(stream)
 }
 
 func testFree(suite *suite.Suite) {
@@ -35,7 +40,7 @@ func testCopyFromToHost(suite *suite.Suite) {
 	test_helpers.ActivateMainDevice()
 	someInts := make([]int32, 1)
 	someInts[0] = 34
-	numBytes := uint(8)
+	numBytes := uint(4)
 	deviceMem, _ := runtime.Malloc(numBytes)
 	deviceMem, err := runtime.CopyToDevice(deviceMem, unsafe.Pointer(&someInts[0]), numBytes)
 	suite.Equal(runtime.Success, err, "Couldn't copy to device due to %v", err)
@@ -44,13 +49,14 @@ func testCopyFromToHost(suite *suite.Suite) {
 	_, err = runtime.CopyFromDevice(unsafe.Pointer(&someInts2[0]), deviceMem, numBytes)
 	suite.Equal(runtime.Success, err, "Couldn't copy to device due to %v", err)
 	suite.Equal(someInts, someInts2, "Elements of host slices do not match. Copying from/to host failed")
+	runtime.Free(deviceMem)
 }
 
 type MemoryTestSuite struct {
 	suite.Suite
 }
 
-func (s *MemoryTestSuite) TestMerkleTree() {
+func (s *MemoryTestSuite) TestMemory() {
 	s.Run("TestMalloc", test_helpers.TestWrapper(&s.Suite, testMalloc))
 	s.Run("TestMallocAsync", test_helpers.TestWrapper(&s.Suite, testMallocAsync))
 	s.Run("TestFree", test_helpers.TestWrapper(&s.Suite, testFree))
